requirements.txt
requests==2.24.0
Flask==1.1.2
flask_login==0.5.0
flask_sqlalchemy==2.4.4
pytest==6.1.0
python-dotenv==0.14.0
redis==3.5.3
rq==1.5.2
selenium==3.141.0
Werkzeug==1.0.1
